+++ tests/cmds/alerts/test_main.py
@@ -33,3 +33,71 @@ def test_send_to(sdk, profile, alert_namespace, mocker, mock_logger_factory, moc
     mock_logger_factory.get_logger_for_server.return_value = logger
     main.send_to(sdk, profile, alert_namespace)
     mock_extract.assert_called_with(sdk, profile, logger, alert_namespace)
+
+
+def test_extract_when_is_advanced_query_and_has_begin_date_exits(sdk, profile, alert_namespace):
+    alert_namespace.advanced_query = "some complex json"
+    alert_namespace.begin = "begin date"
+    with pytest.raises(SystemExit):
+        main.send_to(sdk, profile, alert_namespace)
+
+
+def test_extract_when_is_advanced_query_and_has_end_date_exits(sdk, profile, alert_namespace):
+    alert_namespace.advanced_query = "some complex json"
+    alert_namespace.end = "end date"
+    with pytest.raises(SystemExit):
+        main.print_out(sdk, profile, alert_namespace)
+
+
+@pytest.mark.parametrize(
+    "arg",
+    [
+        "severity",
+        "actor",
+        "actor_contains",
+        "exclude_actor",
+        "exclude_actor_contains",
+        "rule_name",
+        "exclude_rule_name",
+        "rule_id",
+        "exclude_rule_id",
+        "rule_type",
+        "exclude_rule_type",
+    ],
+)
+def test_extract_when_is_advanced_query_and_other_incompatible_multi_narg_argument_passed(
+    sdk, profile, alert_namespace, arg
+):
+    alert_namespace.advanced_query = "some complex json"
+    setattr(alert_namespace, arg, ["test_value"])
+    with pytest.raises(SystemExit):
+        main.write_to(sdk, profile, alert_namespace)
+
+
+@pytest.mark.parametrize("arg", ["state", "description"])
+def test_extract_when_is_advanced_query_and_other_incompatible_single_arg_argument_passed(
+    sdk, profile, alert_namespace, arg
+):
+    alert_namespace.advanced_query = "some complex json"
+    setattr(alert_namespace, arg, "test_value")
+    with pytest.raises(SystemExit):
+        main.print_out(sdk, profile, alert_namespace)
+
+
+def test_extract_when_is_advanced_query_and_use_checkpoint_mode_exits(
+    sdk, profile, alert_namespace
+):
+    alert_namespace.advanced_query = "some complex json"
+    alert_namespace.use_checkpoint = "foo"
+    with pytest.raises(SystemExit):
+        main.print_out(sdk, profile, alert_namespace)
+
+
+def test_extract_when_is_advanced_query_and_does_not_use_checkpoint_does_not_exit(
+    sdk, profile, alert_namespace, mock_extract, mocker, mock_logger_factory
+):
+    logger = mocker.MagicMock()
+    mock_logger_factory.get_logger_for_server.return_value = logger
+    alert_namespace.advanced_query = "some complex json"
+    alert_namespace.use_checkpoint = None
+    main.print_out(sdk, profile, alert_namespace)