+++ tests/test_parser.py
@@ -1,7 +1,13 @@
 import pytest
 
 from code42cli.commands import Command, SubcommandLoader
-from code42cli.parser import ArgumentParserError, CommandParser
+from code42cli.parser import (
+    ArgumentParserError,
+    CommandParser,
+    exit_if_mutually_exclusive_args_used_together,
+)
+
+import code42cli.cmds.search_shared.enums as enums
 
 
 def dummy_method():
@@ -116,3 +122,59 @@ class TestCommandParser(object):
         assert "testsub1" in captured.out
         assert "the subdesc2" in captured.out
         assert "testsub2" in captured.out
+
+
+@pytest.mark.parametrize(
+    "arg",
+    [
+        "c42_username",
+        "actor",
+        "md5",
+        "sha256",
+        "source",
+        "file_name",
+        "file_path",
+        "process_owner",
+        "tab_url",
+    ],
+)
+def test_exit_if_advanced_query_used_with_other_search_args_when_is_advanced_query_and_other_incompatible_multi_narg_argument_passed(
+    file_event_namespace, arg
+):
+    file_event_namespace.advanced_query = "some complex json"
+    setattr(file_event_namespace, arg, ["test_value"])
+    with pytest.raises(SystemExit):
+        exit_if_mutually_exclusive_args_used_together(
+            file_event_namespace, list(enums.FileEventFilterArguments())
+        )
+
+
+def test_exit_if_advanced_query_used_with_other_search_args_when_is_advanced_query_and_uses_checkpoint_does_not_exit_as_invalid_args_does_not_contain_checkpoint(
+    alert_namespace
+):
+    alert_namespace.advanced_query = "some complex json"
+    alert_namespace.use_checkpoint = "foo"
+    exit_if_mutually_exclusive_args_used_together(
+        alert_namespace, list(enums.AlertFilterArguments())
+    )
+
+
+def test_exit_if_advanced_query_used_with_other_search_args_when_is_advanced_query_and_has_include_non_exposure_exits(
+    file_event_namespace
+):
+    file_event_namespace.advanced_query = "some complex json"
+    file_event_namespace.include_non_exposure = True
+    with pytest.raises(SystemExit):
+        exit_if_mutually_exclusive_args_used_together(
+            file_event_namespace, list(enums.FileEventFilterArguments())
+        )
+
+
+def test_exit_if_advanced_query_used_with_other_search_args_when_is_advanced_query_and_has_format_does_not_exit(
+    file_event_namespace
+):
+    file_event_namespace.advanced_query = "some complex json"
+    file_event_namespace.format = "JSON"
+    exit_if_mutually_exclusive_args_used_together(
+        file_event_namespace, list(enums.FileEventFilterArguments())
+    )