+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## 0.7.3 - 2020-06-23
+
+### Fixed
+
+- Fixed bug that caused the last few entries in csv files to sometimes not be processed when performing bulk processing actions.
+
 ## 0.7.2 - 2020-06-11
 
 ### Fixed
@@ -33,6 +39,13 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 - `code42 high-risk-employee remove-risk-tags` now uses required `--username` and `--tag` flags instead of positional arguments.
 
 ### Added
+
+- Extraction subcommands of `code42 security-data`, `print/write-to/send-to` accepts argument `--saved-search` to
+   return saved search results.
+
+- `code42 security-data saved-search` commands:
+    - `list` prints out existing saved searches' id and name
+    - `show` takes a search id
 
 - `code42 high-risk-employee bulk` supports `add-risk-tags` and `remove-risk-tags`.
     - `code42 high-risk-employee bulk generate-template <cmd>` options `add-risk-tags` and `remove-risk-tags`.